// packages/multi-package-json-manager/src/types/tsconfigJson.ts
export type TSconfigJson =
    CompilerOptionsDefinition
    &
    CompileOnSaveDefinition
    &
    TypeAcquisitionDefinition
    &
    ExtendsDefinition
    &
    TsNodeDefinition
    &
    (FilesDefinition | ExcludeDefinition | IncludeDefinition | ReferencesDefinition);
// ...
export interface CompileOnSaveDefinition {
    /**
     * Enable Compile-on-Save for this project.
     */
    compileOnSave?: boolean;

    [ k: string ]: unknown;
}

export interface TypeAcquisitionDefinition {
    /**
     * Auto type (.d.ts) acquisition options for this project. Requires TypeScript version 2.1 or later.
     */
    typeAcquisition?: {
        /**
         * Enable auto type acquisition
         */
        enable?: boolean;
        /**
         * Specifies a list of type declarations to be included in auto type acquisition. Ex. ["jquery", "lodash"]
         */
        include?: string[];
        /**
         * Specifies a list of type declarations to be excluded from auto type acquisition. Ex. ["jquery", "lodash"]
         */
        exclude?: string[];
        [ k: string ]: unknown;
    };

    [ k: string ]: unknown;
}

export interface ExtendsDefinition {
    /**
     * Path to base configuration file to inherit from. Requires TypeScript version 2.1 or later.
     */
    extends?: string;

    [ k: string ]: unknown;
}
// ...
export interface FilesDefinition {
    /**
     * If no 'files' or 'include' property is present in a tsconfig.json, the compiler defaults to including all files in the containing directory and subdirectories except those specified by 'exclude'. When a 'files' property is specified, only those files and those specified by 'include' are included.
     */
    files?: string[];

    [ k: string ]: unknown;
}

export interface ExcludeDefinition {
    /**
     * Specifies a list of files to be excluded from compilation. The 'exclude' property only affects the files included via the 'include' property and not the 'files' property. Glob patterns require TypeScript version 2.0 or later.
     */
    exclude?: string[];

    [ k: string ]: unknown;
}

export interface IncludeDefinition {
    /**
     * Specifies a list of glob patterns that match files to be included in compilation. If no 'files' or 'include' property is present in a tsconfig.json, the compiler defaults to including all files in the containing directory and subdirectories except those specified by 'exclude'. Requires TypeScript version 2.0 or later.
     */
    include?: string[];

    [ k: string ]: unknown;
}

export interface ReferencesDefinition {
    /**
     * Referenced projects. Requires TypeScript version 3.0 or later.
     */
    references?: {
        /**
         * Path to referenced tsconfig or to folder containing tsconfig.
         */
        path?: string;
        [ k: string ]: unknown;
    }[];

    [ k: string ]: unknown;
}